requirements.txt
fastapi
SQLAlchemy
pydantic
uvicorn
databases[aiopg]
asyncpg
alembic
python-multipart
aiofiles
python-jose[cryptography]
passlib[bcrypt]
PyJWT